Se ajusta modelo OperacionCubo para que los valores puedan ser long

+++ Fuentes/Back-End/Model/Models/OperacionCubo.cs
@@ -3,6 +3,6 @@
     public class OperacionCubo
     {
         public string TipoOperacion { get; set; }
-        public int[] InformacionOperacion { get; set; }
+        public long[] InformacionOperacion { get; set; }
     }
 }
